rough pass part 3
done a rough pass of part 3, including a method for creating
gifs and saving the necessary image files from different runs.
non periodic case behaves well. periodic case sometimes still
collapses, need to look into it. but it is definitely periodic
as the collapsed mass will sometimes be on the edge